benchmarks/requirements.txt
msgpack
pytest-benchmark
